Fix bubble chart.
We had an issue that we were only including fragments that have an edit stop
site > than the templates. Sometimes the alignment algorithm can detect the
edit stop site before the stop site of the template. This fixes the issue
however this may or may not be the desired behavior.